requirements.txt
Flask==3.1.1
requests==2.31.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
passlib==1.7.4
mysql-connector-python==8.4.0
